# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Airbnb Visualisation API",
    "description": "API for accessing and analyzing Airbnb listing data",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/analytics/overview": {
      "get": {
        "summary": "Get neighborhood overview statistics",
        "description": "Shows the average cost and density of listings across different London neighborhoods",
        "tags": [
          "Analytics"
        ],
        "responses": {
          "200": {
            "description": "List of neighborhoods with listing count and average price",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "neighbourhood_cleansed": {
                        "type": "string",
                        "description": "Neighborhood name"
                      },
                      "number_of_listings": {
                        "type": "integer",
                        "description": "Number of listings in this neighborhood"
                      },
                      "average_price": {
                        "type": "number",
                        "format": "float",
                        "description": "Average price of listings in this neighborhood"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/room_types": {
      "get": {
        "summary": "Get room type distribution",
        "description": "Provides a breakdown of the types of accommodations available",
        "tags": [
          "Analytics"
        ],
        "responses": {
          "200": {
            "description": "Distribution of listings by room type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "room_type_simple": {
                        "type": "string",
                        "description": "Type of room (e.g., Entire home/apt, Private room)"
                      },
                      "number_of_listings": {
                        "type": "integer",
                        "description": "Number of listings with this room type"
                      },
                      "percentage_of_total": {
                        "type": "number",
                        "format": "float",
                        "description": "Percentage of total listings with this room type"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/room_type_sentiment": {
      "get": {
        "summary": "Get sentiment analysis by room type",
        "description": "Summarizes the proportion of listings by room type that typically receive positive feedback",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "room_type",
            "schema": {
              "type": "string"
            },
            "description": "Filter results by specific room type"
          }
        ],
        "responses": {
          "200": {
            "description": "Sentiment analysis by room type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "room_type": {
                        "type": "string",
                        "description": "Type of room"
                      },
                      "percent_positive_reviews": {
                        "type": "number",
                        "format": "float",
                        "description": "Percentage of positive reviews for this room type"
                      },
                      "number_of_listings": {
                        "type": "integer",
                        "description": "Number of listings with this room type"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/monthly_price": {
      "get": {
        "summary": "Get monthly price trends",
        "description": "Provides an estimated trend of average listing prices over time, grouped by month",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "after",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter results after this date (YYYY-MM-DD format, defaults to 1900-01-01)"
          },
          {
            "in": "query",
            "name": "before",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter results before this date (YYYY-MM-DD format, defaults to 2100-01-01)"
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly price trends over time",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "review_month": {
                        "type": "string",
                        "description": "Month in YYYY-MM format"
                      },
                      "listings_reviewed_count": {
                        "type": "integer",
                        "description": "Number of listings reviewed in this month"
                      },
                      "average_price_of_reviewed_listings": {
                        "type": "number",
                        "format": "float",
                        "description": "Average price of listings reviewed in this month"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid date parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/analytics/hidden_gems": {
      "get": {
        "summary": "Find hidden gem listings",
        "description": "Finds highly rated, high-value listings with relatively few reviews and low price compared to neighbourhood norms",
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "min_rating",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 5
            },
            "description": "Minimum rating threshold for hidden gems (default is 4.8)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of hidden gem listings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "listing_id": {
                        "type": "integer",
                        "description": "Unique identifier of the listing"
                      },
                      "listing_name": {
                        "type": "string",
                        "description": "Name of the listing"
                      },
                      "neighbourhood_cleansed": {
                        "type": "string",
                        "description": "Neighborhood where the listing is located"
                      },
                      "room_type_simple": {
                        "type": "string",
                        "description": "Type of room"
                      },
                      "scores_rating": {
                        "type": "number",
                        "format": "float",
                        "description": "Overall rating score"
                      },
                      "scores_value": {
                        "type": "number",
                        "format": "float",
                        "description": "Value for money rating"
                      },
                      "number_of_reviews": {
                        "type": "integer",
                        "description": "Number of reviews for this listing"
                      },
                      "price": {
                        "type": "number",
                        "format": "float",
                        "description": "Price per night"
                      },
                      "avg_neighbourhood_reviews": {
                        "type": "number",
                        "format": "float",
                        "description": "Average number of reviews for listings in this neighborhood"
                      },
                      "avg_neighbourhood_price_for_room_type": {
                        "type": "number",
                        "format": "float",
                        "description": "Average price for this room type in this neighborhood"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}